requirements.txt
# Django Framework
Django==4.2.9

# Telegram Bot
python-telegram-bot==20.7

# Environment variables
python-dotenv==1.0.0

# HTTP requests (for API polling)
requests==2.31.0

# Date parsing
python-dateutil==2.8.2

# Async support for Django
asgiref==3.7.2

# Production server
gunicorn==21.2.0
